feat :  delete logics about lanes
1. when delete competitions, all related lanes will delete
2. when group is deleted (also delete qualification), then change the qualificationID of lanes to zero

+++ web_server_gin/translate/Competition.go
@@ -190,6 +190,11 @@ func DeleteCompetition(context *gin.Context) {
 			return
 		}
 	}
+	/*delete all related lanes*/
+	success := DeleteLaneByCompetitionId(context, id)
+	if !success {
+		return
+	}
 
 	/*delete competition*/
 	affected, err := database.DeleteCompetition(id)
